Don't set group id if SR_ENV_RUN_TESTS is set
During build of packages which depend on sysrepo, sysrepo is usually
installed for running some tests. These tests may create subscriptions,
push data into sysrepo and read data from sysrepo etc.

Various sysrepo initializing functions such as sr_open, sr_mkpath, call
chown or chmod which the process using sysrepo may not always have the
permissions to do in the build/test env.

This means when a debian package is being built, we will need to run
the unit-tests of the application using sysrepo as root or have
permissions to change file ownership and permissions.

For the principle of least privilige, we shouldn't need to elevate
the process running unit-tests.

In order to achieve this, we check for an environment variable called
`SR_ENV_RUN_TESTS`. If this is set, sysrepo will not call chmod, chown
system calls. In addition to SR_ENV_RUN_TESTS, SR_SHM_PREFIX_ENV is
also expected to be set to a location where the process has permissions
to read and write, such as a temporary dir in /tmp.