sr_cond_futex with counters instead of 2 states
Using a counter as the value of futex prevents unnecessary cpu spinning
when we wake up because of an unlock, but the event we are waiting for
has not completed yet.

With a counter, we can always be sure to not miss
any event, and wait to be notified of a new event.