docs: clarify sr_get_node's return value
I was writing a C++ wrapper around this, and was surprised by what
lyd_path() said about the returned value (instead of the expected
`/example:a/b`, I just got `/example:b`). Of course this makes sense
when the reader is intimately familiar with how libyang works, but it's
better to be explicit about what exactly the implications are here. I
got curious enough to write a test for this, my colleague who reviewed
the bindings were surprised as well, so let's improve the docs.

Since this is probably the only function which returns a "disconnected"
`lyd_node` and there's no explanation of what a "disconnected node"
means, let's just be very succinct here, and improve `lyd_path()`'s docs
in libyang as well.