log error when handler thread exits with failure
Currently when a subscription handler thread exits due to an error, we
do not write any log which could be helpful to diagnose that the
subscription is no longer processing events.

Or, we should not exit the handler thread because there could be
other subscriptions in the context that could still be processed. But
there is no easy way of determining what has failed.
Perhaps, the subscription should be removed but it may not be simple to
convey this to the subscriber.