docs: reserve originator names for RESTCONF (rousette) and sysrepo-cli
No actual originator-specific data are being pushed yet by these
projects, but at least all applications which use sysrepo now have a
list of identifiers to check if they want to see if the originator is
"someone/something talking to a server", in which case the name is one
of the listed ones, or "an internal application".

This is useful for sysrepo-ietf-alarms which exports an RPC that's
"public" for all sysrepo-using applications, but which should not be
accessible via any other channel. It's suboptimal that this is
essentially an ad-hoc list, but it's the easiest tool which we currently
have for this purpose.