sysrepo BUGFIX run_cache realloc wrong array index
When we realloc the run_cache in `sr_conn_run_cache_update`, we first
unlock a read lock and then reacquire a write lock. During this process
another writer could have added more modules to the cache. In this case
the index j would no longer point to the last element of the
`run_cache_mods` array.

So, the new initialization would overwrite the older module, and in
addition leave the newly extended memory uninitialized.