ds json BUGFIX don't rename log files with old notification.
sysrepo uses notification timestamps to rename the log files. The first
timestamp in the name is the timestamp of the first notification, and
the second one is the timestamp of the last notification. However, if
the timestamp is in the past, due to a time jump, the new file name will
be invalid. To avoid losing notification files, don't rename the file
with an older timestamp. The file may contain notification with old
timestamps.

Signed-off-by: Jeremie Leska <jeremie.leska@6wind.com>