tests: race after timeout of high prio done sub
After a timeout of a higher priority subscriber S1, the event is cleared
and the next priority subscribers are notified.
The event and request ID remains the same. Only priority is
changed.

So, S1 can still assume the SHM is valid for it, when it is not.

`sr_shmsub_change_listen_relock` should also check for priority when
determining SHM is still valid, in addition to event and request ID.