shm sub UPDATE orig_cid checks (#3259)
When an event exists, orig_cid must exist. Upon obtaining a new wr_lock
we can clear any previous orig_cid as it is no longer relevant.

This can happen if a previous notifier bailed before waiting
for subscribers to complete. It would clear the event but may not have
been able to clear orig_cid.