shm_sub BUGFIX skip process events when suspended
When a specific subscription in a subscription context is suspended, it
may still contain other active subscriptions.

When the handler thread is woken up to process some active
subscription(s), we check for all available events to process.

However, we must not process events for suspended subscriptions.
If we do, we would be stealing the event from the intended subscribers.

It can also cause asserts to fired when more subs than subscriber_count
have processed the event.

To prevent this, we should also store the suspended state of a
subscription within the subscription context and check it before
processing any events.