scrap and not recover dead subs during notify
During event notification, we only have the ext SHM subscription locked
in READ mode. If we discover that a subscription is dead, we attempted
to recover the SHM by unlocking and acquiring a WRITE lock and
performing cleanup and then reverting to a READ lock.

This was error-prone especially because WRITE locks can become contented
and some return codes had to be ignored.
Further, it needed substantial amount of code to work.

It is better to simply SCRAP these dead subscriptions by marking the
suspended field with a new value(SR_SUBSCR_SCRAPPED=2).
This serves two purposes:
1. For all future event originators, the subscription seems to be
suspended, so no additional code is required there.
2. When another subscription to the same section is added, we check
for SCRAPPED subscriptions and remove them entirely.
This is easy because we already possess a WRITE lock in this case.

A `scraps` flag is set anytime a subscription is marked SCRAPPED.
We search the entire subscription array only if scraps exist.
This should allow subscription addition at minimal additional cost.