shm_sub BUGFIX clear left over event and orig_cid
orig_cid should always be cleared when the originator stops waiting for
the subscribers. If not, subsequent events cannot be generated on the
subscription SHM, until the previous originator process is restarted.

Normally, the wait logic takes care of this. However, in case a
subscription evpipe could not be notified we bail on waiting and return
with error. However, in this case orig_cid and event are left behind
never to be cleared.