test_apply_changes REFACTOR sync with barrier
The callback invocation count is checked in the subscription thread
which is prone to race, when it can be easily checked in the
apply_changes thread.

A single pthread_barrier_wait only ensures that the first
sr_apply_changes has completed.
It does not prevent the second set of sr_apply_changes to start.

This can be synchronized with a double barrier, before and after
checking the invocation count, or by checking the invocation count
in the apply_changes thread, since the `state` is shared by all threads
and the count is atomically manipulated.

This further avoids polling with sleep in the subscribe thread,
simplifying the tests.