dead subscription recovery - review comments
It is not worth the extra code to track scrapped subscriptions in the
SHM segments of each subscription type. This needs significant code to
mark and maintain.

It is easier to simply traverse the entire subscription array before
adding a new subscription of the same type and look for any scraps and
remove them.