sysrepocfg: build-time select default LYD_FORMAT
It is useful to be able to choose the default file format for
sysrepocfg at build time.

For instance, if most users prefer `JSON` over `XML`, it can be
tedious to specify `-f json` for each `sysrepocfg` invocation.

With this change, it will be possible to build sysrepo using
`-DSYSREPOCFG_DEFAULT_LYD_FORMAT=LYD_JSON` and use sysrepocfg.