log when connection is destroyed
A log is emitted when a connection is created, and it is equally useful
to know when it disconnects. This is especially useful to debug any
delays during system shutdown caused by sr_disconnect operations taking
a longer time.

This also helps to debug short-lived processes that use sysrepo by
reading the logs.