remove log for session_start
According to the documentation

> a session is not synchronized at all so it must not be shared among
multiple threads. Each thread should always create its own session
to ensure correct behavior.

and

> Then, an arbitrary number of sessions can be created.
Most importantly, the threading model needs to be considered.
Other than that there are no inherent restrictions in Sysrepo and
because every session requires only little resources,
having many of them should not cause any problems.

However, emitting an INFO level log each time a session is started is
quite noisy. This is especially true for applications which use worker
threads which start sessions and complete the task and end the session.

It doesn't seem to be very useful to have this log and processes that
want to log this can explicitly print it as all the info is available
through API.