tests BUGFIX test_mult_update sync threads
One of the threads can starve the other unless we make sure that after
each iteration, the other thread is also given a chance to run.
Also, for the purposes of the test, it is preferable if both threads
are running in parallel, instead of one thread being starved.

Without this, valgrind test_apply_changes sometimes fail, where only one
thread runs for an extended period of time.