ds redis, ds mongo UPDATE gap between uo elements
Gap between elements of user-ordered lists or leaf-lists should be a
power of two. Elements added at the start of the list/leaf-list
obtain a value of the first element in that list divided by two.
Therefore a maximum of 9 insertions (log2(1000)) at the start of the
list/leaf-list was possible without the need to shift the other
elements. Now 10 insertions are possible (log2(1024)) without shifting.