remove scrap subscription logic for dead subs
It is sufficient to check for dead subscriptions before adding a new
subscription. It is not necessary to mark the dead subscriptions found
during event notification.

The advantages of marking dead subscriptions as scrapped were:
1. avoid redundant connection live checks.
2. avoid waking up oper poll threads often when an oper get sub dies.

However, big code changes are needed to enable this and the
performance impact is not considered to be significant.