shm ext UPDATE do not recover dead subs during notify (#3332)
* scrap and not recover dead subs during notify

During event notification, we only have the ext SHM subscription locked
in READ mode. If we discover that a subscription is dead, we attempted
to recover the SHM by unlocking and acquiring a WRITE lock and
performing cleanup and then reverting to a READ lock.

This was error-prone especially because WRITE locks can become contented
and some return codes had to be ignored.
Further, it needed substantial amount of code to work.

It is better to simply SCRAP these dead subscriptions by marking the
suspended field with a new value(SR_SUBSCR_SCRAPPED=2).
This serves two purposes:
1. For all future event originators, the subscription seems to be
suspended, so no additional code is required there.
2. When another subscription to the same section is added, we check
for SCRAPPED subscriptions and remove them entirely.
This is easy because we already possess a WRITE lock in this case.

A `scraps` flag is set anytime a subscription is marked SCRAPPED.
We search the entire subscription array only if scraps exist.
This should allow subscription addition at minimal additional cost.

* dead subscription recovery - review comments

It is not worth the extra code to track scrapped subscriptions in the
SHM segments of each subscription type. This needs significant code to
mark and maintain.

It is easier to simply traverse the entire subscription array before
adding a new subscription of the same type and look for any scraps and
remove them.

* test for subscription scraps and recovery

Add tests for subscription scrapping and recovery is performed when
a new subscription is created in the same module and datastore.

* remove scrap subscription logic for dead subs

It is sufficient to check for dead subscriptions before adding a new
subscription. It is not necessary to mark the dead subscriptions found
during event notification.

The advantages of marking dead subscriptions as scrapped were:
1. avoid redundant connection live checks.
2. avoid waking up oper poll threads often when an oper get sub dies.

However, big code changes are needed to enable this and the
performance impact is not considered to be significant.

* notif-sub check dead cid before evpipe notify

We should always check and avoid attempting to notify dead cids
as when evpipe notify fails, the event is considered a failure.

* shm_ext add missing docs for dead subscr removal

The 4 new functions for ext SHM removal of dead subscriptions were
missing doxygen comments.

1. sr_shmext_change_sub_remove_dead
2. sr_shmext_oper_get_sub_remove_dead
3. sr_shmext_oper_poll_sub_remove_dead
4. sr_shmext_notif_sub_remove_dead