call getenv only once per process
sysrepo does not handle any changes to environmental variables that it
reads during the lifetime of the process.

We should use a static variable to read the env vars first time and
cache the values to minimize the risk any application changes them.

SR_PATH_MAX is defined as 256 and is defined as the maximum length of
runtime configurable paths for sysrepo.