bugfix strncpy null term of maxlen SR_PATH_MAX
`strncpy` does not result in a destination string but a character
sequence. If 'dst' is not long enough to contain the null byte,
'dst' will not be null-terminated.

`stlcpy` does this but is not portable.

`strncpy` expects us to pass the size of the destination buffer as the
limiting size for copying.

It also expects that the source string fits into the destination
buffer including the null-termination byte, if the result is to be a
string and not a character sequence.

The strlen() function calculates the length of the string pointed
to by s, excluding the terminating null byte ('\0').

So, the correct way to use `strncpy()` is:

char dst[DST_SZ];
char src[SRC_SZ];
...
if (strnlen(src, SRC_SZ) < DST_SZ) {
    strncpy(dst, src, DST_SZ);
}

See `String vs character sequence` in manpage of `string_copying`.
strlcpy or strscpy maybe alternatives, but with other challenges such as
portability or availability.