shm_sub ensure recovery on previous incomplete cleanup
A process holding a subscription could die for any number of reasons and
not always have the opportunity to clean up completely.

Sometimes, we may end up with a incompletely cleaned subscription/SHM
where the shm file is deleted but the subscription is not fully removed
yet.

In this case, we should ensure the subscription SHM is recovered before
adding a new subscription.

However, by checking the return value of `unlink` and not making
allowance for the shm file to be missing, we may be forever stuck, being
unable to add a new subscription.

For example, the below failure occurs when the process is restarted:

```
[app::sysrepo] Recovering module "ietf-interfaces" operational get subscription of CID 11.
[app::sysrepo] Failed to unlink "/dev/shm/srsub_ietf-interfaces.oper.f05e29e6" SHM (No such file or directory).
[app::session] sr_oper_get_subscribe() failed: 3
```

So, we should make an exception for `errno == ENOENT` while checking
unlink return code.