check sub is_alive before copying running to oper
While fetching operational data, data from the running datastore is
copied automatically if there is a non passive subscription to the
running datastore for that module.

But it is also possible that this subscription is no longer alive, in
which case we should not be reporting this data in the operational
datastore.