ds_mongo support multiple discard-items
sysrepo discard-items are stored with the key /sysrepo:discard-items.
This key (_id) should be unique.

But unlike other nodes, multiple sysrepo discard-items nodes can be
present.

So, we should also add an index to make these nodes unique. This index
is added with a prefix 'O' for easy detection.

While loading data back, look for the prefix 'O' to detect opaque keys
and discard the prefix and index to get the original path back.