sr_module_oper_data_load - shorten read ext_lock
We don't need to hold onto ext_lock for the entire duration of
sr_module_oper_data_load.

We only need it to get the sr_mod_oper_push_t array for the module.

So quickly make a copy of it and unlock, instead of holding onto ext SHM
lock while we load the operational data.

Modules with no push data can be skipped without acquiring ext_lock.