shm ext_hole_lock can be mutex instead of rwlock
The `ext_hole_lock` does not need to be a rwlock since it is always
acquired in WRITE mode.

It can be acquired in READ mode for `sr_shmext_print()` calls during
deletions, but is most likely not worth the extra code and effort as
`sr_shmext_print()` returns early in release builds.