ext_lock is not needed for oper push READ or DEL
The system-wide ext_lock is needed only when
1. ext SHM needs to be remapped or truncated
2. ext SHM holes are added/deleted/merged

Reading oper push data and deleting the oper push data from the mod shm
does not cause either of the above two operations.

So, system-wide ext_lock is not needed.

The process specific (among different threads of the same process) lock
`conn->ext_remap_lock` already ensures that reading the ext_shm is safe
and the current process will not remap to learn new changes from other
processes.

The appropriate module locks that are already held, ensure that the ext
SHM locations do not change while being read.

This means that
1. sr_module_oper_data_load
2. sr_shmext_oper_push_get
3. sr_shmmod_del_module_oper_data

do not need to acquire the system-wide ext_lock.