shm_ext - add locks for reading suspended state
No lock was obtained while reading the suspended state of subscriptions.
The dedicated READ lock must be obtained to prevent data races.
For instance, another process may remove the data concurrently.

- do not need WRITE remap lock to remove subscriptions even for
`sr_shmext_FOO_sub_check()`. a READ ext remap lock is enough.

- for sub_check() functions, also remember when all locks are unlocked
and we fail to obtain a WRITE lock. We should not try to unlock.
A correct goto cleanup is used, but add this to be clear.

- fixup comments for the ext lock being acquired and released.
explicitly use the term `CONN EXT REMAP` for the connection specific
rwlock that guards various sessions of the same connection.

The term `EXT LOCK` only refers to the process-shared (across multiple
connections) mutex from now.

When ext_remap_lock is acquired in WRITE mode along with ext_lock, it is
referred to as `REMAP(WRITE) AND EXT LOCK`.