sr_shmext_print determine validate internally
It is not required to pass an additional validate argument to
`sr_shmext_print()`. Since it is always called when `ext_lock` mutex
is held, It is only necessary to check if we have mapped the latest
`Ext SHM`, by comparing the actual file size vs mapped size.

If we have the latest file mapped, we can validate.