shm_ext BUGFIX sr_shmext__sub_suspended locks
All the sr_shmext_FOO_sub_suspended() functions acquire a WRITE module
subscription lock to change the subscription state, but no lock for
reading the subscription suspended state.

They need to acquire a READ module subscriptions lock to read safely.
Although the data field itself is atomic, without the READ lock, there
is no guarantee that the location in Ext SHM being read from refers to
this subscription. For example, It could be moved due to addition of
another subscription to the same module.