shm_sub BUGFIX update sub_count for DONE ABORT ev
sr_shmsub_change_notify_evpipe() updates the subscriber_count with
the number of subscribers actually notified.

This updated count should be written into the subscription SHM for all
change_sub events. Previously only CHANGE and UPDATE events accounted
for this. However, this must be handled correctly for DONE and ABORT
events as well.

This allows for avoiding unnecessary waiting for "dead" subscriptions
which are discovered late during evpipe notify.