oper ds plugins - skip load and create diff on del
None of the operational datastore plugins work with a diff but actual
data to be stored.

So, it is unnecessary to load the data and create a diff for storage.

A common scenario would be when a process with large operational data
restarts abruptly, and when it comes back up, it has to clean up the
previous data. This can be very CPU intensive and is easily avoided.