shm_ext BUGFIX oper push change of order
The code in `sr_shmext_oper_push_update_get_idx()` and
`sr_shmext_oper_push_update()` is complex and corrupts the Ext SHM if
the order changes in some ways.

To cover all scenarios, it is better to simply use `qsort()` to keep
the oper push data of a given module sorted correctly.

This simplifies the logic significantly. There might be some overhead
due to sorting, but should not be significant as
1. Order is changed very infrequently.
2. It is unlikely that too many sessions push operational data to same
   module.