oper_push UPDATE order set and get
changing oper push order of a session's module data when some data
exists can result in changes in the operational datastore, if the module
data has any 'discard-items' nodes pushed by any session.

These changes are not notified when the order is changed.

To simplify and keep things straight-forward, it is now forbidden to
change the order if the session has oper push data.
It is also forbidden to change the order for all modules as it is not
trivial to ensure correctness. For instance, an sr_apply_changes() call
may fail at a later stage, as another session has that order reserved.