ds plugins OPTIMIZE pass req xpaths to load_cb
load_cb can be xpath aware for operational datastore, and fetch only the
required data for the datastore plugins that support it.

Implemented support:
1. redis DS plugin supports this for all datastores.
2. mongo ds plugin supports this except for operational datastore.
  - ietf-origin metadata must be explicitly retrieved for each path.
  - this is not implemented yet, so all data will be retrieved.
3. json ds plugin - unimplemented. ignores passed xpaths.

This has the benefit of optimally loading only the requested data, and
not constructing the entire module datatree, if only a small part is
being read.

Further, it can reduce the amount of time modules remain read locked,
when they have large amount of data, and occasional readers of small
data.