ntf_json UPDATE return if notification file space is insufficient
When there is not enough space at a particular time stamp for
registering the notification. We'll try to re-create the same file,
resulting in a 'File exists' error.
If the file doesn't have enough space, check timestamps, and if they
are within the exact same second, clean up and set a proper error
message.

Signed-off-by: Nelson Lako Tangala <nelson.lako_tangala@6wind.com>