compat UPDATE use __atomic instead of legacy __sync func (#3595)
According to gcc documentation:

These functions are implemented in terms of the `__atomic` builtins
(see Built-in Functions for Memory Model Aware Atomic Operations).
They should not be used for new code which should use the
`__atomic` builtins instead.

These functions are intended to replace the legacy ‘__sync’ builtins.
The main difference is that the memory order that is requested is a
parameter to the functions.
New code should always use the ‘__atomic’ builtins rather than
the ‘__sync’ builtins.

The previous simple implementations for ATOMIC_LOAD_RELAXED() and
ATOMIC_STORE_RELAXED() not always be atomic.