change_sub OPTIMIZE WRITE event only if needed
When change subscribers are notified at a given priority, there
is a bottleneck of the sub_shm SR_LOCK_WRITE_URGE lock, for each
subscriber that processes the event, in order to update the progress of
the event processing.

Specifically, the following items are updated as relevant:
1. subscriber_count - remaining subscribers yet to process the event.
2. if there is data to be sent back (for an UPDATE/RPC event etc.)
3. event errors.
4. if there are no remaining subscribers, then final result.

The most common of these is scenario 1.
It is possible to use atomic operations to update `subscriber_count`,
while holding the sub_shm READ lock.

This allows for event processing to continue in parallel without
stopping all other subscribers, each time a subscriber finishes
processing.

For the final subscriber to process the event, the WRITE lock is
readily available, as there are no other readers.