sr_rwunlock BUGFIX broadcast if readers has 1 spot
Once the reader list is full, and one reader unlocks, the last
spot becomes available for a new reader. This condition must be
broadcast, in case there are any waiters for a READ lock.

So, check if the last but one
readers[SR_RWLOCK_READ_LIMIT-2] exists,
and the last one
readers[SR_RWLOCK_READ_LIMIT-1] is empty.

Without this, readers waiting after a full readers list, will not be
woken up until all readers unlock.