modinfo BUGFIX duplicate reader in srmon ds_lock
There is a data race with unlocked access of rwlock data for populating
sysrepo monitoring module lock data in
`sr_modinfo_module_srmon_locks_ds()`.

To accurately report locks in sysrepo monitoring, The only correct
solution would be to quickly cache the entire readers array with the
rwlock mutex held and unlock, and then construct the sysrepo-monitoring
tree using the cached data.

But even this comes with the problem of having to acquire and release a
lot of locks, if there are a lot of modules, and is not ideal.
Also, if any lock is WRITE locked, then it leads to delays.

So, the simplest solution is to ignore duplicate readers
(which can happen rarely) by passing the `LYD_NEW_PATH_UPDATE`
flag to `sr_lyd_new_path()`.