rwlock BUGFIX sr_rwlock_reader_recover
 ### Brief
simplify the complex memmove logic for reader_del to keep readers and
read_count arrays consistent.

 ### Problem description
Although memmove() is MT-safe, it is not an atomic operation.
Also if the process crashes after memmove() of `readers` but before
memmove() `read_count`, the rwlock is not recoverable.

 ### Solution
Swap the last reader with the reader being deleted. This means that
there are only 2 swaps necessary to complete the reader deletion.

With this, there is only one data race possible:
The last reader CID can be present in duplicates, at the location of the
just deleted reader and at the end, concurrently.

But this is easily recoverable as implemented in
sr_rwlock_reader_recover().

 ### Example:
reader_cid(read_count) [2(1), 7(1), 4(1), 5(3)]

CID 7 unlocks with the following steps:

Step 1. decr read_count
  [2(1), 7(0), 4(1), 5(3)]

Step 2. store last CID read_count under CID 7
  [2(1), 7(3), 4(1), 5(3)]

Step 3. change CID 7 to CID 5
  [2(1), 5(3), 4(1), 5(3)]

Step 4. zero last CID
  [2(1), 5(3), 4(1), 0(3)]

Step 5. zero last CID count
  [2(1), 5(9), 4(1), 0(0)]

Specifically a duplicate reader at the end of the readers array is
present if process CID 7 dies just after Step 3. So, as part of
recovering the rwlock, we should also check for this condition and
handle it properly.

Also, it is important to maintain the correct order while deleting the
CID from the readers list.
For example, if reader CID is updated before read_count, we may be left
with
  [2(1), 5(0), 4(1), 5(3)]
after a crash, which is not easy to recover as CID 5 is alive, but it's
read_count is reported as 0.