add tests for rwlock reader limit and recovery
There are no tests covering the scenarios where a rwlock readers list is
full.

If the readers array gets corrupted, it can make the system unusable.
Adding these tests now to cover this gap.

also add a test for sr_unsubscribe retry on first lock acq failure if
the callback thread was busy.