sysrepo UPDATE schema mount rework
Store schema mount data in a file, whenever its change is detected. This
data is then parsed, whenever mod_info is used, which guarantees that
the ly_ctx used to write the data and the ly_ctx used to parse the data
are the same (ly_ctx is locked everytime before modinfo init).