sr_lycc_store_context write to filename not fd
Context lock has become significantly more complex when a printed
context is in use with caches.
The caches in question are:
1. running datastore cache
2. oper poll subscription cache
3. oper push data cache (in the sessions).

We need to track all the users of the three caches.

The main problem is that caches held in all processes and connections
need to be flushed before the context can change.
There is only one way to flush all the caches - have a
`flush_cache_rpc` and a thread to process it.

 ## Printed context
Printed context is written into `sr_ctx` SHM file, and accessed via a
file descriptor. The same filename is used by every process to directly
reuse the printed context.

 ## Context switch
A system-wide WRITE context lock is acquired, stopping all other
operations. The open file descriptor is overwritten with new content.
So, the old context immediately disappears for all processes, and old
caches can no longer be flushed.

 ## Solution
To prevent this, the new context is written into a temporary SHM file
and renamed to `sr_ctx`. The old fd (and old context) is still available
to all other processes. When a new context lock is acquired, other
processes detect the changed context, and first flush their caches, and
then close the old context fd, and open the new `sr_ctx` file.

Once all processes close the fd for the old context file, it will be
removed from memory by the operating system.