add enabled module change sub with less locks
Adding a change sub can be done with much lesser locking.

1. do not hold module READ lock after calling the enabled callback.
=> check for any missed events and process them again.

2. do not hold change sub WRITE lock during the enabled callback.
=> enabled callback can take a long time, and holding change sub WRITE
lock means that sr_apply_changes() cannot even prepare the notify diff.

3. do not hold subs WRITE lock along with change sub WRITE lock.
=> This allows event processing to continue for other subs handled by
this subscription context.

After being added to Ext SHM, subscriber must process new events.
A notification sent before the subscription was added to the
subscription context structure, could have been missed.
So, we should check if there is a pending event, and notify the event
pipe again for processing.

Any DONE events generated before the ENABLED callback is invoked,
should be skipped, because the ENABLED callback already informs the
subscriber of this changeset. Only the callback invocation must be
skipped, and the event processed normally as the publisher expects it to
be processed.

 ### Summary of locking for adding change sub with these changes:
1. change sub WRITE lock is held only for adding sub to Ext SHM.
2. module READ lock is held only for invoking the enabled callback.
3. subs WRITE lock is held only for adding sub to the subscription
   context.

- tests add a test to check data races.
It's difficult to reproduce the races without adding some sleep in
sr_module_change_subscribe() after CHANGE SUB WRITE UNLOCK.